fix hubstorage writer bug triggered when flushing buffer I hit this bug after attempting to use the Hubstorage Writer with a bigger dataset, testing Datahub internal request: https://staging.scrapinghub.com/p/28865/job/1/6844/#/log?line=1298 Seems it had been tested only with small ones. :)
The problem is that previous write_batch implementation was calling
parent write_batch to update item_count in metadata, but the parent
method was designed to work with temp file buffers, so it would try
to write a buffer file.

I updated the implementation to conform with current expectations
for write_batch (and to leverage HubstorageClient batch upload and
compression), but I couldn't help but think that BaseWriter is kinda
doing too much -- perhaps it makes sense to move the temp buffer
behavior to FilebaseBasedWriter or something in between.