Retry reading streams
Now the stream structure only has file name and stream size and
optionally any metadata (needs to be serializable) that the reader might
need to open the stream. Streams are only opened when necessary (might
not be opened if we are restarting a export) and may  be opened several
times if we fail to read all data from the stream.